+++ relanz/config/settings.py
@@ -39,8 +39,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'user',
-    'account',
     'community',
+    # 'challenge',
 ]
 
 MIDDLEWARE = [
