+++ relanz/templates/main/home.html
@@ -12,10 +12,10 @@
     {% if user.is_authenticated %}
         <a href="{% url 'user:signout' %}">로그아웃</a>
         <h1>환영합니다! {{user.username}} 님</h1>
-        {% if not account.nickname %}
-        <a href="{% url 'account:nickname' %}">프로필 설정</a>
+        {% if not user.nickname %}
+        <a href="{% url 'user:nickname' %}">프로필 설정</a>
         {% else %}
-        <a href="{% url 'account:accountInfo' account.id %}">프로필 보기</a>
+        <a href="{% url 'user:accountInfo' user.id %}">프로필 보기</a>
         {% endif %}
     {% else %}
         <a href="{% url 'user:signup' %}">회원가입</a>
