+++ relanz/templates/user/content.html
@@ -19,7 +19,7 @@
     <input type="radio" name="sex" value="female">
     <label>여자</label>
 
-    <button class="accountInfo" type="submit" action="{% url 'account:content' %}">다음</button>
+    <button class="accountInfo" type="submit" action="{% url 'user:content' %}">다음</button>
     <br>
 </form>
