+++ relanz/user/models.py
@@ -1,10 +1,17 @@
 from django.db import models
+from django.conf import settings
 from django.contrib.auth.models import AbstractUser
 
 # Create your models here.
 
 class User(AbstractUser):
-    password = models.CharField(max_length=20, blank=False)
-    created_at = models.DateField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
+    nickname = models.CharField(max_length=32, unique=True, blank=False)
+    birth = models.IntegerField(null=True, default=0)
+    
+    sex_tuple = [
+        ('male', 'male'),
+        ('female', 'female')
+    ]
+
+    sex = models.CharField(max_length=10, choices=sex_tuple, blank=False)
