+++ relanz/user/views.py
@@ -1,6 +1,5 @@
 from django.shortcuts import render, redirect
 from .models import User
-from account.models import Account
 from django.contrib.auth import login, logout, authenticate
 
 # Create your views here.
@@ -16,13 +15,13 @@ def signup(request):
 
         if not (username and password and re_password):
             res_data['error']="입력되지 않은 값이 있습니다."
-            render(request, 'user/signup.html', res_data)
+            return render(request, 'user/signup.html', res_data)
         elif (password != re_password):
             res_data['error']="비밀번호가 일치하지 않습니다."
-            render(request, 'user/signup.html', res_data)
+            return render(request, 'user/signup.html', res_data)
         elif User.objects.filter(username=username):
             res_data['error']="이미 존재하는 아이디입니다."
-            render(request, 'user/signup.html', res_data)
+            return render(request, 'user/signup.html', res_data)
         else:
             user=User.objects.create_user(
                 username=username,
@@ -50,3 +49,50 @@ def signin(request):
 def signout(request):
     logout(request)
     return render(request, 'user/signout.html')
+
+
+
+
+def nickname(request):
+    user_data={}
+    if request.method=="GET":
+        return render(request, 'user/nickname.html')
+    if request.method=="POST":
+        nickname = request.POST.get('nickname')
+        user = request.user
+        
+        if not nickname:
+            user_data['error'] = "닉네임을 입력해주세요."
+            return render(request, 'user/nickname.html', user_data)
+        
+        elif User.objects.filter(nickname=nickname).exists():
+            user_data['error'] = "이미 존재하는 닉네임입니다."
+            return render(request, 'user/nickname.html', user_data)
+        
+        else:
+            user = User(nickname=nickname)
+            user.user = user
+            user.save()
+            return redirect('user:content')
+
+
+
+def content(request):
+    if request.method=="GET":
+        user = User(user=request.user)
+        return render(request, 'user/content.html', {'user':user})
+    if request.method=="POST":
+        user = User.objects.get(user=request.user)
+        birth = request.POST.get('birth')
+        sex = request.POST.get('sex')
+        user.birth = birth
+        user.sex = sex
+        user.save()
+        return redirect('user:accountInfo', user.id)
+    
+
+def accountInfo(request, user_id):
+    user = request.user
+    user = User.objects.get(user)
+    user.birth = 2023 - user.birth
+    return render(request, 'user/accountInfo.html', {'user':user})
